+++ Questionable/Controller/Steps/Shared/Gather.cs
@@ -40,7 +40,6 @@ internal static class Gather
     internal sealed class DelayedGatheringExecutor
     (
         GatheringPointRegistry gatheringPointRegistry,
-        TerritoryData territoryData,
         IClientState clientState,
         IObjectTable objectTable,
         IServiceProvider serviceProvider,
@@ -95,7 +94,7 @@ internal static class Gather
 
             uint territoryId = gatheringRoot.Steps.Last().TerritoryId;
             yield return new WaitCondition.Task(() => clientState.TerritoryType == territoryId,
-                $"Wait(territory: {territoryData.GetNameAndId(territoryId)})");
+                $"Wait(territory: {TerritoryData.GetNameAndId(territoryId)})");
 
             yield return new WaitNavmesh.Task();