+++ Questionable/Windows/QuestComponents/CreationUtilsComponent.cs
@@ -37,7 +37,6 @@ internal sealed class CreationUtilsComponent
     QuestRegistry questRegistry,
     QuestFunctions questFunctions,
     CameraFunctions cameraFunctions,
-    TerritoryData territoryData,
     QuestData questData,
     QuestSelectionWindow questSelectionWindow,
     PriorityWindow priorityWindow,
@@ -56,7 +55,7 @@ internal sealed class CreationUtilsComponent
         if (objectTable[0] == null)
             return;
 
-        string territoryName = territoryData.GetNameAndId(clientState.TerritoryType);
+        string territoryName = TerritoryData.GetNameAndId(clientState.TerritoryType);
         ImGui.Text(territoryName);
 
         if (gameFunctions.IsFlyingUnlockedInCurrentZone())
@@ -280,7 +279,7 @@ internal sealed class CreationUtilsComponent
             bool showZoneQuests = ImGuiComponentsLocal.IconButton(FontAwesomeIcon.MapMarkerAlt);
             if (showZoneQuests)
             {
-                logger.LogDebug($"打开当前区域的任务选择窗口 {territoryData.GetNameAndId(clientState.TerritoryType)}");
+                logger.LogDebug($"打开当前区域的任务选择窗口 {TerritoryData.GetNameAndId(clientState.TerritoryType)}");
                 questSelectionWindow.OpenForCurrentZone();
             }
             if (ImGui.IsItemHovered())
@@ -355,11 +354,7 @@ internal sealed class CreationUtilsComponent
                 };
                 ImGui.SetClipboardText($$"""
                                          "DataId": {{GameFunctions.GetBaseID(target)}},
-                                                   "Position": {
-                                                     "X": {{target.Position.X.ToString(CultureInfo.InvariantCulture)}},
-                                                     "Y": {{target.Position.Y.ToString(CultureInfo.InvariantCulture)}},
-                                                     "Z": {{target.Position.Z.ToString(CultureInfo.InvariantCulture)}}
-                                                   },
+                                                   "Position": {{target.Position.ToJsonString()}},
                                                    "TerritoryId": {{clientState.TerritoryType}},
 
                                          """ + (GameFunctions.IsFlyingUnlocked(clientState.TerritoryType) ?