+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+- Feature: Skip aetheryte shortcut if in target territory -alydev
+
 - Feature: "Stuck?" button is now more visible when QST is running -alydev
 - Feature: Troubleshooting info now includes player job/level/position/territory(location)
+  - Bug fix: i broke the button, but it's working again now